CB-238 feat : modify searchPage to ProductPage Component - 따로있던 검색 페이지를 상품페이지 내부의 컴포넌트로 이동
